% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_date_rightalign.R
\name{breaks_right}
\alias{breaks_right}
\title{Generate date breaks that align with the maximum data value}
\usage{
breaks_right(limits, n_breaks = 5, ...)
}
\arguments{
\item{limits}{Length-two numeric or date vector}

\item{n_breaks}{Number of breaks; passed to \code{scales::breaks_pretty()}}

\item{...}{Passed to \code{scales::breaks_pretty()}}
}
\value{
Vector of breaks, of the same class as \code{limits}
}
\description{
This function generates a vector of of 'pretty'
breaks (using \code{scales::breaks_pretty()}) that ends with
the upper limit provided and excludes any values that lie
outside the limits.
}
\examples{
# Can be used with numeric vectors
breaks_right(c(10, 30))

# Or date vectors
econ_dates <- range(ggplot2::economics$date)
breaks_right(econ_dates)

# Can be supplied directly to the `breaks` argument of
# `ggplot2::scale_*_continuous()`. Note that the breaks will
# be aligned to the right-limit of the plot
# (including expansion/padding area).

library(ggplot2)

ggplot(ggplot2::economics,
      aes(x = date, y = unemploy)) +
 geom_line() +
 scale_x_date(breaks = breaks_right)

# To right-align to the data, not including padding, try:
ggplot(ggplot2::economics,
    aes(x = date, y = unemploy)) +
    geom_line() +
    scale_x_date(breaks = breaks_right(limits = range(
    ggplot2::economics$date)))

# Or use `scale_x_date_rightalign()`:
ggplot(ggplot2::economics,
    aes(x = date, y = unemploy)) +
    geom_line() +
    scale_x_date_rightalign()

}
\seealso{
scale_x_date_rightalign()
}
\author{
Originally written by Matt Cowgill for the \code{djprtheme} package,
in which the Department of Jobs, Precincts and Regions (Victoria) is
the copyright holder.
}
