% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_finallabel.R, R/geom_richlegend.R
\docType{data}
\name{geom_finallabel}
\alias{geom_finallabel}
\alias{GeomFinalLabel}
\alias{GeomRichLegend}
\title{Line with a dot on the final observation}
\format{
An object of class \code{GeomFinalLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 7.

An object of class \code{GeomRichLegend} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 7.
}
\usage{
geom_finallabel(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  nudge_x = 0,
  nudge_y = 0,
  nudge_x_perc = 1.5,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

GeomFinalLabel

GeomRichLegend
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix (e.g. \code{"count"} rather than
\code{"stat_count"})}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function. Cannot be jointy specified with
\code{nudge_x} or \code{nudge_y}.}

\item{nudge_x, nudge_y}{Horizontal and vertical adjustment to nudge labels by.
Useful for offsetting text from points, particularly on discrete scales.
Cannot be jointly specified with \code{position}.}

\item{nudge_x_perc}{Amount to nudge label along the x-axis, expressed as a
percentage of the data range along the x-axis. The default is \code{1.5}, which
will nudge the text 1.5\% of the data range from the final point. This is applied
in addition to \code{nudge_x}, which nudges the text a specific number of data units.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
\code{geom_finallabel()} draws a \code{ggplot2::geom_text()}
at the observation with the maximum x value for each group.
}
\section{Aesthetics}{

\code{geom_text()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{x}}
\item \strong{\code{y}}
\item \strong{\code{label}}
\item \code{alpha}
\item \code{angle}
\item \code{colour}
\item \code{family}
\item \code{fontface}
\item \code{group}
\item \code{hjust}
\item \code{lineheight}
\item \code{size}
\item \code{vjust}
}

Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}

\examples{
library(ggplot2)

ggplot(ggplot2::economics_long, aes(x = date, y = value)) +
  geom_linepoint(aes(col = variable)) +
  geom_finallabel(aes(label = value)) +
  facet_wrap(~variable)
}
\seealso{
\code{ggplot2::geom_text}
}
\keyword{datasets}
