% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_richlegend.R
\name{rt_draw_panel}
\alias{rt_draw_panel}
\title{Slightly modified from gridtext by Claus O Wilke}
\usage{
rt_draw_panel(
  data,
  panel_params,
  coord,
  label.padding = unit(c(0.25, 0.25, 0.25, 0.25), "lines"),
  label.margin = unit(c(0, 0, 0, 0), "lines"),
  label.r = unit(0.15, "lines"),
  na.rm = FALSE
)
}
\description{
Slightly modified from gridtext by Claus O Wilke
}
